requirements.txt
Cython==0.20.1
gevent==1.0
greenlet==0.4.2
gsocketpool==0.1.2
hash-ring==1.3.1
msgpack-python==0.4.1
nose==1.3.0
wsgiref==0.1.2
